Move all data files into data direcory. Prepare for data rework,
