+++ packages/kartoffelgames.web.potato_web_builder/source/core/data/module-values.ts
@@ -1,4 +1,4 @@
-import { ScopedValues } from '../scoped-values';
+import { ScopedValues } from './scoped-values';
 import { ModuleValueProcedure } from './module-value-procedure';
 
 /**
