+++ packages/kartoffelgames.web.potato_web_builder/source/core/data/scoped-values.ts
@@ -1,5 +1,5 @@
 import { Dictionary, Exception } from '@kartoffelgames/core';
-import { Component } from './component/component';
+import { Component } from '../component/component';
 
 /**
  * Interface between persistent values directly from component and temporary values
