+++ packages/kartoffelgames.web.potato_web_builder/source/core/module/instruction_module/instruction-result.ts
@@ -1,6 +1,6 @@
 import { Exception, List } from '@kartoffelgames/core';
 import { PwbTemplate } from '../../component/template/nodes/pwb-template';
-import { ScopedValues } from '../../scoped-values';
+import { ScopedValues } from '../../data/scoped-values';
 
 /**
  * Results for instruction module.
