+++ packages/kartoffelgames.web.potato_web_builder/source/module/pwb_child/pwb-child.decorator.ts
@@ -1,8 +1,8 @@
 import { Exception } from '@kartoffelgames/core';
 import { Component, ComponentProcessor } from '../../core/component/component';
 import { ComponentRegister } from '../../core/component/component-register';
-import { ScopedValues } from '../../core/scoped-values';
-import { ComponentScopedValues } from '../../core/component/injection_reference/component-scoped-values';
+import { ScopedValues } from '../../core/data/scoped-values';
+import { ComponentScopedValues } from '../../core/data/component-scoped-values';
 
 /**
  * AtScript. Id child
