+++ packages/kartoffelgames.web.potato_web_builder/test/integration/component/scoped-values.test.ts
@@ -1,10 +1,10 @@
 import { expect } from 'chai';
 import { ComponentRegister } from '../../../source/core/component/component-register';
-import { ComponentScopedValues } from '../../../source/core/component/injection_reference/component-scoped-values';
+import { ComponentScopedValues } from '../../../source/core/data/component-scoped-values';
 import { PwbComponent } from '../../../source/core/component/pwb-component.decorator';
 import { PwbConfiguration } from '../../../source/core/configuration/pwb-configuration';
 import { Processor } from '../../../source/core/core_entity/processor';
-import { ScopedValues } from '../../../source/core/scoped-values';
+import { ScopedValues } from '../../../source/core/data/scoped-values';
 import '../../mock/request-animation-frame-mock-session';
 import '../../utility/chai-helper';
 import { TestUtil } from '../../utility/test-util';
